requirements.txt
asgiref==3.5.0
Django==4.0.3
django-ckeditor==6.3.2
django-js-asset==2.0.0
djangorestframework==3.13.1
gunicorn==20.1.0
Pillow==9.0.1
psycopg2==2.9.3
pytz==2021.3
sqlparse==0.4.2
